ECF-3 make:auth

+++ src/Controller/EmprunteurController.php
@@ -3,12 +3,16 @@
 namespace App\Controller;
 
 use App\Entity\Emprunteur;
+use App\Entity\User;
 use App\Form\EmprunteurType;
 use App\Repository\EmprunteurRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
+
 
 /**
  * @Route("/emprunteur")
@@ -35,11 +39,19 @@ class EmprunteurController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $user = $emprunteur->getUser();
+            $user->setPassword(
+                $passwordEncoder->encodePassword(
+                    $user,
+                    $form->get('user')->get('plainPassword')->getData()
+                )
+            );
+
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->persist($emprunteur);
             $entityManager->flush();
 
-            return $this->redirectToRoute('emprunteur_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('emprunteur_index');
         }
 
         return $this->render('emprunteur/new.html.twig', [
@@ -51,8 +63,15 @@ class EmprunteurController extends AbstractController
     /**
      * @Route("/{id}", name="emprunteur_show", methods={"GET"})
      */
-    public function show(Emprunteur $emprunteur): Response
-    {
+    public function show(Emprunteur $emprunteur, EmprunteurRepository $emprunteurRepository): Response
+    {       
+        
+        $response = $this->redirectEmprunteur('emprunteur_show', $emprunteur, $emprunteurRepository);
+
+        if ($response) {
+            return $response;
+        }
+
         return $this->render('emprunteur/show.html.twig', [
             'emprunteur' => $emprunteur,
         ]);
@@ -61,15 +80,30 @@ class EmprunteurController extends AbstractController
     /**
      * @Route("/{id}/edit", name="emprunteur_edit", methods={"GET","POST"})
      */
-    public function edit(Request $request, Emprunteur $emprunteur): Response
+    public function edit(Request $request, Emprunteur $emprunteur, EmprunteurRepository $emprunteurRepository, UserPasswordEncoderInterface $passwordEncoder): Response
     {
+        $response = $this->redirectEmprunteur('emprunteur_edit', $emprunteur, $emprunteurRepository);
+
+        if ($response) {
+            return $response;
+        }
+
         $form = $this->createForm(EmprunteurType::class, $emprunteur);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            // encode the plain password
+            $user = $emprunteur->getUser();
+            $user->setPassword(
+                $passwordEncoder->encodePassword(
+                    $user,
+                    $form->get('user')->get('plainPassword')->getData()
+                )
+            );
+
             $this->getDoctrine()->getManager()->flush();
 
-            return $this->redirectToRoute('emprunteur_index', [], Response::HTTP_SEE_OTHER);
+            return $this->redirectToRoute('emprunteur_index');
         }
 
         return $this->render('emprunteur/edit.html.twig', [
@@ -83,12 +117,40 @@ class EmprunteurController extends AbstractController
      */
     public function delete(Request $request, Emprunteur $emprunteur): Response
     {
+        if ($this->isGranted('ROLE_EMPRUNTEUR')) {
+            throw new AccessDeniedException();
+        }
+
         if ($this->isCsrfTokenValid('delete'.$emprunteur->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($emprunteur);
             $entityManager->flush();
         }
 
-        return $this->redirectToRoute('emprunteur_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('emprunteur_index');
+    }
+
+    private function redirectEmprunteur(string $route, Emprunteur $emprunteur, EmprunteurRepository $emprunteurRepository)
+    {
+        // Récupération du compte de l'utilisateur qui est connecté
+        $user = $this->getUser();
+
+        // On vérifie si l'utilisateur est un emprunteur 
+        if (in_array('ROLE_EMPRUNTEUR', $user->getRoles())) {
+            // if ($this->isGranted('ROLE_EMPRUNTEUR')) {
+                // Récupèration du profil emprunteur
+            $userEmprunteur = $emprunteurRepository->findOneByUser($user);
+
+            // Comparaison du profil demandé par l'utilisateur et le profil de l'utilisateur
+            // Si les deux sont différents, on redirige l'utilisateur vers la page de son profil
+            if ($emprunteur->getId() != $userEmprunteur->getId()) {
+                return $this->redirectToRoute($route, [
+                    'id' => $userEmprunteur->getId(),
+                ]);
+            }
+        }
+
+        // Si aucune redirection n'est nécessaire, on renvoit une valeur nulle
+        return null;
     }
 }
